code/requirements.txt
PyYAML>=3.10-11
requests>=2.14.2
